+++ src/main/java/org/springframework/data/couchbase/config/AbstractReactiveCouchbaseConfiguration.java
@@ -13,120 +13,114 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.springframework.data.couchbase.config;
 
-import java.util.HashSet;
-import java.util.Set;
+import java.util.List;
 
 import com.couchbase.client.java.Bucket;
+import com.couchbase.client.java.Cluster;
+import com.couchbase.client.java.CouchbaseCluster;
 import com.couchbase.client.java.cluster.ClusterInfo;
-import org.springframework.beans.factory.config.BeanDefinition;
+import com.couchbase.client.java.env.CouchbaseEnvironment;
+import com.couchbase.client.java.env.DefaultCouchbaseEnvironment;
+
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ClassPathScanningCandidateComponentProvider;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.type.filter.AnnotationTypeFilter;
-import org.springframework.data.annotation.Persistent;
-import org.springframework.data.couchbase.core.RxJavaCouchbaseTemplate;
-import org.springframework.data.couchbase.core.mapping.Document;
-import org.springframework.data.couchbase.repository.ReactiveCouchbaseRepository;
-import org.springframework.data.couchbase.repository.config.ReactiveRepositoryOperationsMapping;
-import org.springframework.util.ClassUtils;
-import org.springframework.util.StringUtils;
 
 /**
- * Base class for Reactive Spring Data Couchbase configuration using JavaConfig.
+ *
+ * Base class for Reactive Spring Data Couchbase configuration java config
  *
  * @author Subhashni Balakrishnan
  */
 @Configuration
-public abstract class AbstractReactiveCouchbaseConfiguration extends CouchbaseConfigurationSupport {
+public abstract class AbstractReactiveCouchbaseConfiguration
+        extends AbstractReactiveCouchbaseDataConfiguration implements CouchbaseConfigurer {
 
-	@Override
-	protected CouchbaseConfigurer couchbaseConfigurer() {
-		return this;
-	}
+    /**
+     * The list of hostnames (or IP addresses) to bootstrap from.
+     *
+     * @return the list of bootstrap hosts.
+     */
+    protected abstract List<String> getBootstrapHosts();
 
-	/**
-	 * Creates a {@link ReactiveCouchbaseTemplate}.
-	 *
-	 * This uses {@link #mappingCouchbaseConverter()}, {@link #translationService()} and {@link #getDefaultConsistency()}
-	 * for construction.
-	 *
-	 * Additionally, it will expect injection of a {@link ClusterInfo} and a {@link Bucket} beans from the context (most
-	 * probably from another configuration). For a self-sufficient configuration that defines such beans, see
-	 * {@link CouchbaseConfigurationSupport}.
-	 *
-	 * @throws Exception on Bean construction failure.
-	 */
-	@Bean(name = BeanNames.RXJAVA1_COUCHBASE_TEMPLATE)
-	public RxJavaCouchbaseTemplate reactiveCouchbaseTemplate() throws Exception {
-		RxJavaCouchbaseTemplate template = new RxJavaCouchbaseTemplate(couchbaseConfigurer().couchbaseClusterInfo(),
-				couchbaseConfigurer().couchbaseClient(), mappingCouchbaseConverter(), translationService());
-		template.setDefaultConsistency(getDefaultConsistency());
-		return template;
-	}
+    /**
+     * The name of the bucket to connect to.
+     *
+     * @return the name of the bucket.
+     */
+    protected abstract String getBucketName();
 
-	/**
-	 * Creates the {@link ReactiveRepositoryOperationsMapping} bean which will be used by the framework to choose which
-	 * {@link ReactiveCouchbaseOperations} should back which {@link ReactiveCouchbaseRepository}.
-	 * Override {@link #configureReactiveRepositoryOperationsMapping} in order to customize this.
-	 *
-	 * @throws Exception
-	 */
-	@Bean(name = BeanNames.REACTIVE_COUCHBASE_OPERATIONS_MAPPING)
-	public ReactiveRepositoryOperationsMapping reactiveRepositoryOperationsMapping(RxJavaCouchbaseTemplate couchbaseTemplate) throws  Exception {
-		//create a base mapping that associates all repositories to the default template
-		ReactiveRepositoryOperationsMapping baseMapping = new ReactiveRepositoryOperationsMapping(couchbaseTemplate);
-		//let the user tune it
-		configureReactiveRepositoryOperationsMapping(baseMapping);
-		return baseMapping;
-	}
+    /**
+     * The password of the bucket (can be an empty string).
+     *
+     * @return the password of the bucket.
+     */
+    protected abstract String getBucketPassword();
 
+    /**
+     * Is the {@link #getEnvironment()} to be destroyed by Spring?
+     *
+     * @return true if Spring should destroy the environment with the context, false otherwise.
+     */
+    protected boolean isEnvironmentManagedBySpring() {
+        return true;
+    }
 
-	/**
-	 * In order to customize the mapping between repositories/entity types to couchbase templates,
-	 * use the provided mapping's api (eg. in order to have different buckets backing different repositories).
-	 *
-	 * @param mapping the default mapping (will associate all repositories to the default template).
-	 */
-	protected void configureReactiveRepositoryOperationsMapping(ReactiveRepositoryOperationsMapping mapping) {
-		//NO_OP
-	}
+    /**
+     * Override this method if you want a customized {@link CouchbaseEnvironment}.
+     * This environment will be managed by Spring, which will call its shutdown()
+     * method upon bean destruction, unless you override {@link #isEnvironmentManagedBySpring()}
+     * as well to return false.
+     *
+     * @return a customized environment, defaults to a {@link DefaultCouchbaseEnvironment}.
+     */
+    protected CouchbaseEnvironment getEnvironment() {
+        return DefaultCouchbaseEnvironment.create();
+    }
 
+    @Override
+    protected CouchbaseConfigurer couchbaseConfigurer() {
+        return this;
+    }
 
-	/**
-	 * Scans the mapping base package for classes annotated with {@link Document}.
-	 *
-	 * @throws ClassNotFoundException if initial entity sets could not be loaded.
-	 */
-	@Override
-	protected Set<Class<?>> getInitialEntitySet() throws ClassNotFoundException {
-		String basePackage = getMappingBasePackage();
-		Set<Class<?>> initialEntitySet = new HashSet<Class<?>>();
+    @Override
+    @Bean(destroyMethod = "shutdown", name = BeanNames.COUCHBASE_ENV)
+    public CouchbaseEnvironment couchbaseEnvironment() {
+        CouchbaseEnvironment env = getEnvironment();
+        if (isEnvironmentManagedBySpring()) {
+            return env;
+        }
+        return new CouchbaseEnvironmentNoShutdownProxy(env);
+    }
 
-		if (StringUtils.hasText(basePackage)) {
-			ClassPathScanningCandidateComponentProvider componentProvider = new ClassPathScanningCandidateComponentProvider(false);
-			componentProvider.addIncludeFilter(new AnnotationTypeFilter(Document.class));
-			componentProvider.addIncludeFilter(new AnnotationTypeFilter(Persistent.class));
-			for (BeanDefinition candidate : componentProvider.findCandidateComponents(basePackage)) {
-				initialEntitySet.add(ClassUtils.forName(candidate.getBeanClassName(), AbstractReactiveCouchbaseConfiguration.class.getClassLoader()));
-			}
-		}
-		return initialEntitySet;
-	}
+    /**
+     * Returns the {@link Cluster} instance to connect to.
+     *
+     * @throws Exception on Bean construction failure.
+     */
+    @Override
+    @Bean(destroyMethod = "disconnect", name = BeanNames.COUCHBASE_CLUSTER)
+    public Cluster couchbaseCluster() throws Exception {
+        return CouchbaseCluster.create(couchbaseEnvironment(), getBootstrapHosts());
+    }
 
-	/**
-	 * Return the base package to scan for mapped {@link Document}s. Will return the package name of the configuration
-	 * class (the concrete class, not this one here) by default.
-	 * <p/>
-	 * <p>So if you have a {@code com.acme.AppConfig} extending {@link AbstractCouchbaseConfiguration} the base package
-	 * will be considered {@code com.acme} unless the method is overridden to implement alternate behavior.</p>
-	 *
-	 * @return the base package to scan for mapped {@link Document} classes or {@literal null} to not enable scanning for
-	 * entities.
-	 */
-	protected String getMappingBasePackage() {
-		return getClass().getPackage().getName();
-	}
+    @Override
+    @Bean(name = BeanNames.COUCHBASE_CLUSTER_INFO)
+    public ClusterInfo couchbaseClusterInfo() throws Exception {
+        return couchbaseCluster().clusterManager(getBucketName(), getBucketPassword()).info();
+    }
 
+    /**
+     * Return the {@link Bucket} instance to connect to.
+     *
+     * @throws Exception on Bean construction failure.
+     */
+    @Override
+    @Bean(destroyMethod = "close", name = BeanNames.COUCHBASE_BUCKET)
+    public Bucket couchbaseClient() throws Exception {
+        //@Bean method can use another @Bean method in the same @Configuration by directly invoking it
+        return couchbaseCluster().openBucket(getBucketName(), getBucketPassword());
+    }
 }
